Verify the purchase transaction and validade the amount transferred. Change prices to decimal type.

+++ Business/Advisor/EscrowResultBusiness.cs
@@ -44,7 +44,7 @@ namespace Auctus.Business.Advisor
                     var endValue = history.Where(c => c.Date == purchase.ExpirationDate.Value).Single().RealValue;
                     var performedPercentage = (endValue / startValue) - 1.0;
 
-                    double buyerTokenAmount = 0, ownerTokenAmount = 0;
+                    decimal buyerTokenAmount = 0, ownerTokenAmount = 0;
                     if (estimatedPercentage > performedPercentage)
                         buyerTokenAmount = purchase.Price;
                     else
